chore: move faker into own file
